requirements.txt
langchain>=1.2.10
langchain-community>=0.3.0
langchain-unstructured>=0.1.6
langchain-ollama>=1.0.0
langchain-openai>=1.0.3
langchain-qdrant>=1.1.0
langchain-text-splitters>=1.0.0
unstructured>=0.15.0
pdfplumber>=0.11.0
python-docx>=1.1.2
Pillow>=10.4.0
pydantic>=2.12,<3
python-dotenv>=1.0.1
qdrant-client>=1.15.1
fastapi>=0.115.0
uvicorn>=0.34.0
python-multipart>=0.0.20
jinja2>=3.1.0
pytest>=7.0.0